requirements.txt
fastapi
uvicorn
pillow
python-dotenv
